+++ packages/workers-auth/tests/access.test.ts
@@ -1,20 +1,55 @@
-import ci from "ci-info";
-import { beforeEach, describe, it, vi } from "vitest";
+import { setupServer } from "msw/node";
+import {
+	afterAll,
+	afterEach,
+	beforeAll,
+	beforeEach,
+	describe,
+	it,
+	vi,
+} from "vitest";
 import {
 	clearAccessCaches,
 	domainUsesAccess,
 	getAccessHeaders,
-} from "../user/access";
-import { mockConsoleMethods } from "./helpers/mock-console";
-import { useMockIsTTY } from "./helpers/mock-istty";
-import { msw, mswAccessHandlers } from "./helpers/msw";
+} from "../src/access";
+import { mswAccessHandlers } from "../src/test-helpers/msw-handlers/access";
+
+vi.mock("node:child_process", () => ({
+	spawnSync: vi.fn((binary: string) => {
+		if (binary === "cloudflared") {
+			return { error: true };
+		}
+		return {
+			error: null,
+			stdout: Buffer.from(""),
+			stderr: Buffer.from(""),
+			status: 0,
+		};
+	}),
+}));
+
+const msw = setupServer();
+
+beforeAll(() => msw.listen({ onUnhandledRequest: "error" }));
+afterEach(() => {
+	msw.restoreHandlers();
+	msw.resetHandlers();
+});
+afterAll(() => msw.close());
+
+const silentLogger = {
+	debug: () => {},
+	info: () => {},
+	log: () => {},
+	warn: vi.fn(),
+	error: () => {},
+};
 
 describe("access", () => {
-	const { setIsTTY } = useMockIsTTY();
-	const std = mockConsoleMethods();
-
 	beforeEach(() => {
 		clearAccessCaches();
+		silentLogger.warn = vi.fn();
 		msw.use(...mswAccessHandlers);
 	});
 
@@ -55,13 +90,15 @@ describe("access", () => {
 				vi.stubEnv("CLOUDFLARE_ACCESS_CLIENT_ID", "test-client-id.access");
 				vi.stubEnv("CLOUDFLARE_ACCESS_CLIENT_SECRET", "test-client-secret");
 
-				const headers = await getAccessHeaders("access-protected.com");
+				const headers = await getAccessHeaders("access-protected.com", {
+					logger: silentLogger,
+				});
 				expect(headers).toEqual({
 					"CF-Access-Client-Id": "test-client-id.access",
 					"CF-Access-Client-Secret": "test-client-secret",
 				});
 				// No warning is presented since both env variables are set
-				expect(std.warn).toMatchInlineSnapshot(`""`);
+				expect(silentLogger.warn).not.toHaveBeenCalled();
 			});
 
 			it("should return service token headers for a service-auth-only domain (403 response)", async ({
@@ -71,37 +108,43 @@ describe("access", () => {
 				// only allow Service Auth tokens, the domain responds with a
 				// hard 403 instead of redirecting to cloudflareaccess.com.
 				// `domainUsesAccess` returns false in this case, so the env var
-				// check must happen first - otherwise Wrangler would return
-				// empty headers and the request would fail with a 403.
+				// check must happen first - otherwise empty headers would be
+				// returned and the request would fail with a 403.
 				vi.stubEnv("CLOUDFLARE_ACCESS_CLIENT_ID", "test-client-id.access");
 				vi.stubEnv("CLOUDFLARE_ACCESS_CLIENT_SECRET", "test-client-secret");
 
-				const headers = await getAccessHeaders("access-service-auth-only.com");
+				const headers = await getAccessHeaders("access-service-auth-only.com", {
+					logger: silentLogger,
+				});
 				expect(headers).toEqual({
 					"CF-Access-Client-Id": "test-client-id.access",
 					"CF-Access-Client-Secret": "test-client-secret",
 				});
-				expect(std.warn).toMatchInlineSnapshot(`""`);
+				expect(silentLogger.warn).not.toHaveBeenCalled();
 			});
 
 			it("should warn when only CLOUDFLARE_ACCESS_CLIENT_ID is set", async ({
 				expect,
 			}) => {
 				vi.stubEnv("CLOUDFLARE_ACCESS_CLIENT_ID", "test-client-id.access");
-				setIsTTY(false);
 
 				await expect(
-					getAccessHeaders("access-protected.com")
+					getAccessHeaders("access-protected.com", {
+						logger: silentLogger,
+						isNonInteractiveOrCI: () => true,
+					})
 				).rejects.toThrowErrorMatchingInlineSnapshot(
 					`[Error: The domain "access-protected.com" is behind Cloudflare Access, but no Access Service Token credentials were found and the current environment is non-interactive.
 Set the CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET environment variables to authenticate with an Access Service Token.
 See https://developers.cloudflare.com/cloudflare-one/access-controls/service-credentials/service-tokens/]`
 				);
-				expect(std.warn).toContain(
-					"Both CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET must be set"
+				expect(silentLogger.warn).toHaveBeenCalledWith(
+					expect.stringContaining(
+						"Both CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET must be set"
+					)
 				);
-				expect(std.warn).toContain(
-					"Only CLOUDFLARE_ACCESS_CLIENT_ID was found"
+				expect(silentLogger.warn).toHaveBeenCalledWith(
+					expect.stringContaining("Only CLOUDFLARE_ACCESS_CLIENT_ID was found")
 				);
 			});
 
@@ -109,20 +152,26 @@ See https://developers.cloudflare.com/cloudflare-one/access-controls/service-cre
 				expect,
 			}) => {
 				vi.stubEnv("CLOUDFLARE_ACCESS_CLIENT_SECRET", "test-client-secret");
-				setIsTTY(false);
 
 				await expect(
-					getAccessHeaders("access-protected.com")
+					getAccessHeaders("access-protected.com", {
+						logger: silentLogger,
+						isNonInteractiveOrCI: () => true,
+					})
 				).rejects.toThrowErrorMatchingInlineSnapshot(
 					`[Error: The domain "access-protected.com" is behind Cloudflare Access, but no Access Service Token credentials were found and the current environment is non-interactive.
 Set the CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET environment variables to authenticate with an Access Service Token.
 See https://developers.cloudflare.com/cloudflare-one/access-controls/service-credentials/service-tokens/]`
 				);
-				expect(std.warn).toContain(
-					"Both CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET must be set"
+				expect(silentLogger.warn).toHaveBeenCalledWith(
+					expect.stringContaining(
+						"Both CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET must be set"
+					)
 				);
-				expect(std.warn).toContain(
-					"Only CLOUDFLARE_ACCESS_CLIENT_SECRET was found"
+				expect(silentLogger.warn).toHaveBeenCalledWith(
+					expect.stringContaining(
+						"Only CLOUDFLARE_ACCESS_CLIENT_SECRET was found"
+					)
 				);
 			});
 		});
@@ -131,25 +180,10 @@ See https://developers.cloudflare.com/cloudflare-one/access-controls/service-cre
 			it("should throw actionable error when non-interactive and no service token", async ({
 				expect,
 			}) => {
-				setIsTTY(false);
-
 				await expect(
-					getAccessHeaders("access-protected.com")
-				).rejects.toThrowErrorMatchingInlineSnapshot(
-					`[Error: The domain "access-protected.com" is behind Cloudflare Access, but no Access Service Token credentials were found and the current environment is non-interactive.
-Set the CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET environment variables to authenticate with an Access Service Token.
-See https://developers.cloudflare.com/cloudflare-one/access-controls/service-credentials/service-tokens/]`
-				);
-			});
-
-			it("should throw actionable error when in CI and no service token", async ({
-				expect,
-			}) => {
-				setIsTTY(true);
-				vi.mocked(ci).isCI = true;
-
-				await expect(
-					getAccessHeaders("access-protected.com")
+					getAccessHeaders("access-protected.com", {
+						isNonInteractiveOrCI: () => true,
+					})
 				).rejects.toThrowErrorMatchingInlineSnapshot(
 					`[Error: The domain "access-protected.com" is behind Cloudflare Access, but no Access Service Token credentials were found and the current environment is non-interactive.
 Set the CLOUDFLARE_ACCESS_CLIENT_ID and CLOUDFLARE_ACCESS_CLIENT_SECRET environment variables to authenticate with an Access Service Token.
@@ -162,11 +196,10 @@ See https://developers.cloudflare.com/cloudflare-one/access-controls/service-cre
 			it("should error without cloudflared installed on an access protected domain", async ({
 				expect,
 			}) => {
-				setIsTTY(true);
-				vi.mocked(ci).isCI = false;
-
 				await expect(
-					getAccessHeaders("access-protected.com")
+					getAccessHeaders("access-protected.com", {
+						isNonInteractiveOrCI: () => false,
+					})
 				).rejects.toThrowErrorMatchingInlineSnapshot(
 					`[Error: To use Wrangler with Cloudflare Access, please install \`cloudflared\` from https://developers.cloudflare.com/cloudflare-one/connections/connect-apps/install-and-setup/installation]`
 				);
