+++ packages/workers-auth/tests/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"extends": "@cloudflare/workers-tsconfig/tsconfig.json",
+	"compilerOptions": {
+		"module": "esnext",
+		"types": ["node"],
+		"noEmit": true
+	},
+	"include": ["**/*.ts", "../src/**/*.ts"]
+}
