+++ packages/workers-auth/tsup.config.ts
@@ -1,0 +1,20 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig(() => [
+	{
+		treeshake: true,
+		keepNames: true,
+		entry: ["src/index.ts", "src/test-helpers/index.ts"],
+		platform: "node",
+		format: "esm",
+		dts: true,
+		outDir: "dist",
+		tsconfig: "tsconfig.json",
+		metafile: true,
+		sourcemap: process.env.SOURCEMAPS !== "false",
+		define: {
+			"process.env.NODE_ENV": `'${"production"}'`,
+		},
+		external: ["@cloudflare/*", "vitest", "undici", "msw"],
+	},
+]);
