+++ packages/workers-auth/turbo.json
@@ -1,0 +1,16 @@
+{
+	"$schema": "http://turbo.build/schema.json",
+	"extends": ["//"],
+	"tasks": {
+		"build": {
+			"inputs": ["$TURBO_DEFAULT$", "!**/tests/**"],
+			"outputs": ["dist/**"],
+			"env": ["SOURCEMAPS"],
+			"passThroughEnv": ["PWD"]
+		},
+		"test:ci": {
+			"dependsOn": ["build"],
+			"env": ["LC_ALL", "TZ"]
+		}
+	}
+}
