+++ packages/workers-auth/vitest.config.mts
@@ -1,0 +1,15 @@
+import path from "node:path";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+	test: {
+		testTimeout: 15_000,
+		pool: "forks",
+		include: ["**/tests/**/*.test.ts"],
+		globalSetup: path.resolve(__dirname, "tests/vitest.global.ts"),
+		setupFiles: [path.resolve(__dirname, "tests/vitest.setup.ts")],
+		reporters: ["default"],
+		unstubEnvs: true,
+		mockReset: true,
+	},
+});
