+++ packages/wrangler/src/user/generate-auth-url.ts
@@ -1,33 +1,7 @@
-interface GenerateAuthUrlProps {
-	authUrl: string;
-	clientId: string;
-	scopes: string[];
-	stateQueryParam: string;
-	codeChallenge: string;
-}
-
-export const OAUTH_CALLBACK_URL = "http://localhost:8976/oauth/callback";
-
-/**
- * generateAuthUrl was extracted from getAuthURL in user.tsx
- * to make it possible to mock the generated URL
- */
-export const generateAuthUrl = ({
-	authUrl,
-	clientId,
-	scopes,
-	stateQueryParam,
-	codeChallenge,
-}: GenerateAuthUrlProps) => {
-	return (
-		authUrl +
-		`?response_type=code&` +
-		`client_id=${encodeURIComponent(clientId)}&` +
-		`redirect_uri=${encodeURIComponent(OAUTH_CALLBACK_URL)}&` +
-		// we add offline_access manually for every request
-		`scope=${encodeURIComponent([...scopes, "offline_access"].join(" "))}&` +
-		`state=${stateQueryParam}&` +
-		`code_challenge=${encodeURIComponent(codeChallenge)}&` +
-		`code_challenge_method=S256`
-	);
-};
+// Re-export shim. The real implementation lives in `@cloudflare/workers-auth`.
+//
+// This file exists so wrangler's tests can continue to `vi.mock("../user/generate-auth-url", ...)`
+// to produce deterministic OAuth URLs for snapshot testing. The mocked exports
+// are imported by `./oauth.ts` and injected into the OAuth flow context,
+// where the workers-auth package uses them internally.
+export { generateAuthUrl, OAUTH_CALLBACK_URL } from "@cloudflare/workers-auth";
